Fix 10: adminCRUD fixes
adminCRUD had some legacy design that should be aligned with other parts
of the app. The call for account information is filtered in the backend
and is in need of a new function.

The design issues are taken care of. All Svg components resides in one
big component instead of being scattered. The account table has a new
design that looks better. Old buttons are replaced with new Svg
clickable components.