feat: Dockerize frontend
The backend has been dockerized and now it's time for the frontend to be dockerized too.

In this commit, environment variables have been established in different files. Depending on which mode the server starts with, a different variable will be used. logIn.jsx has been used for testing that communication works.